refactor: 선택기능 관련 로직 useSelectedItem으로 분리
